chore(quill): extract the presets panel, retire the composer name
DateRangePresetsPanel gets its own module (selection types, chip
vocabularies, range helpers, panel JSX), leaving date-time-picker.tsx
with just the mode wiring. Composer naming is gone: the insight host is
InsightQuillDateFilter, panel data-attrs are date-presets-*, app footer
rows are date-filter-*, with lemon-skin selectors updated in step. The
flag key keeps its name — renaming it would orphan the existing flag.
