+++ frontend/src/styles/lemon-skin.scss
@@ -16,11 +16,11 @@
  * Quill's own rules live in `@layer components`, so these unlayered overrides
  * win the cascade without specificity hacks.
  *
- * Covered: only what the first consumer (the quill date composer,
- * InsightDateFilterComposer) renders — Button, Input, InputGroup, Switch,
+ * Covered: only what the first consumer (the quill date filter,
+ * InsightQuillDateFilter) renders — Button, Input, InputGroup, Switch,
  * Badge (→ LemonTag), Separator, Label, Popover/select surfaces,
  * ToggleGroup (→ Lemon day chips), the DateTimePicker footer, and the
- * composer itself — plus a core-token rebind that turns bespoke Tailwind
+ * presets panel itself — plus a core-token rebind that turns bespoke Tailwind
  * in quill-based features lemony wholesale. A fuller prototype covering
  * every primitive family existed before this trim; grow sections back as
  * new surfaces adopt the skin.
@@ -29,7 +29,7 @@
  * renders in portals at <body>, outside any scoped wrapper. Two ways to
  * cover it: put the attributes on <body>/<html> (the eventual app-wide flip
  * deployment), or pass `data-lemon-skin` to the quill *Content component so
- * it lands on the portaled element itself — the composer's flyout and unit
+ * it lands on the portaled element itself — the exclusions flyout and unit
  * dropdown do the latter, and the portal-surface rules match both placements.
  *
  * Known gaps, acceptable for a prototype:
@@ -422,7 +422,7 @@
 
 /* Chips → LemonTag metrics: flat neutral border, tag-scale type — 16 of Lemon's
    raised secondary buttons in one grid read as noise. */
-[data-lemon-skin] .quill-button:is([data-attr^='date-composer-short-'], [data-attr^='date-composer-fixed-']) {
+[data-lemon-skin] .quill-button:is([data-attr^='date-presets-chip-']) {
     height: auto;
     padding: 0.125rem 0.25rem;
     font-size: 0.6875rem;
@@ -436,10 +436,7 @@
     translate: none;
 }
 
-[data-lemon-skin]
-    .quill-button:is([data-attr^='date-composer-short-'], [data-attr^='date-composer-fixed-']):not(
-        [aria-selected='true']
-    ):hover {
+[data-lemon-skin] .quill-button:is([data-attr^='date-presets-chip-']):not([aria-selected='true']):hover {
     background: var(--lemon-skin-surface);
     border-color: var(--lemon-skin-border-hover);
     box-shadow: none;
@@ -447,8 +444,7 @@
 }
 
 /* Selected chip → LemonTag highlight: accent text + border, no fill. */
-[data-lemon-skin]
-    .quill-button:is([data-attr^='date-composer-short-'], [data-attr^='date-composer-fixed-'])[aria-selected='true'] {
+[data-lemon-skin] .quill-button:is([data-attr^='date-presets-chip-'])[aria-selected='true'] {
     color: var(--lemon-skin-accent);
     background: var(--lemon-skin-surface);
     border-color: var(--lemon-skin-accent);
@@ -456,7 +452,7 @@
 
 /* Footer rows (Custom range… / Exclude) → Lemon menu-row metrics.
    `height`, not `min-height`: the size-sm variant sets an explicit height that would win. */
-[data-lemon-skin] .quill-button:is([data-attr='date-composer-custom-range'], [data-attr='date-composer-exclusions']) {
+[data-lemon-skin] .quill-button:is([data-attr='date-presets-custom-range'], [data-attr='date-filter-exclusions']) {
     height: 2.0625rem;
     padding-inline: 0.5rem;
     font-size: 0.8125rem;
@@ -474,7 +470,7 @@
 
 /* Exact-time row label → same metrics as the footer button rows (the generic
    Label→LemonLabel mapping above is 600/14px, too heavy for a menu row). */
-[data-lemon-skin] label[for='composer-exact-time'] {
+[data-lemon-skin] label[for='date-filter-exact-time'] {
     font-size: 0.8125rem;
     font-weight: 500;
 }
